backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pyjwt==2.9.0
httpx==0.27.2
python-multipart
python-magic
slowapi
pdfminer.six
python-docx
python-jose[cryptography]
pydantic-settings>=2.0.0
pypdf
faster-whisper
requests==2.32.3